refactor(core): create core/ sub-package — 8 files (slice 6/8, #170)
