+++ src/voicecli/adapters/nats/_tts_runner.py
@@ -107,7 +107,7 @@ async def run_synthesis(
             # pulling torch when only inspecting the adapter (e.g. --help).
             from voicecli.adapters.synthesis import LocalSynthesisAdapter  # noqa: PLC0415
             from voicecli.runtime.model_registry import model_registry  # noqa: PLC0415
-            from voicecli.utils import UNRESTRICTED  # noqa: PLC0415
+            from voicecli.core.utils import UNRESTRICTED  # noqa: PLC0415
 
             kw = dict(optional_kwargs)
             if language is not None:
