+++ src/voicecli/core/__init__.py
@@ -1,0 +1,1 @@
+# voicecli.core — configuration, paths, utilities, models, samples, STT modes, history
