+++ src/voicecli/core/paths.py
