+++ src/voicecli/core/stt_modes.py
