+++ src/voicecli/engines/chatterbox.py
@@ -7,8 +7,8 @@ from typing import TYPE_CHECKING
 
 from voicecli.engines.engine import cuda_guard
 from voicecli.engines._chatterbox_base import ChatterboxBase
-from voicecli.models import CHATTERBOX_MODEL, warn_if_first_download
-from voicecli.utils import resolve_language as _resolve_language
+from voicecli.core.models import CHATTERBOX_MODEL, warn_if_first_download
+from voicecli.core.utils import resolve_language as _resolve_language
 
 if TYPE_CHECKING:
     from voicecli.api.markdown import Segment
