+++ src/voicecli/engines/chatterbox_turbo.py
@@ -7,7 +7,7 @@ from typing import TYPE_CHECKING
 
 from voicecli.engines.engine import cuda_guard
 from voicecli.engines._chatterbox_base import ChatterboxBase
-from voicecli.models import CHATTERBOX_MODEL, warn_if_first_download
+from voicecli.core.models import CHATTERBOX_MODEL, warn_if_first_download
 
 if TYPE_CHECKING:
     from voicecli.api.markdown import Segment
