+++ tests/test_output_path_validation.py
@@ -6,7 +6,7 @@ from unittest.mock import MagicMock, patch
 import pytest
 
 from voicecli.api import _validate_output_path, clone, generate
-from voicecli.utils import OUTPUT_DIR, UNRESTRICTED
+from voicecli.core.utils import OUTPUT_DIR, UNRESTRICTED
 
 
 class TestValidateOutputPath:
@@ -77,7 +77,7 @@ class TestValidateOutputPath:
         # Redirect OUTPUT_DIR → tmp_path so the test is hermetic.
         fake_base = tmp_path / "voices_out"
         fake_base.mkdir()
-        monkeypatch.setattr("voicecli.api.OUTPUT_DIR", fake_base)
+        monkeypatch.setattr("voicecli.api.api.OUTPUT_DIR", fake_base)
         inside = fake_base / "x.wav"
         result = _validate_output_path(inside, allowed_base=fake_base)
         assert result == inside.resolve()
@@ -97,8 +97,8 @@ class TestGenerateOutputValidation:
 
         # Act + Assert
         with (
-            patch("voicecli.config.load_defaults", return_value={}),
-            patch("voicecli.engine.get_engine", side_effect=Exception("should not reach")),
+            patch("voicecli.core.config.load_defaults", return_value={}),
+            patch("voicecli.engines.engine.get_engine", side_effect=Exception("should not reach")),
             pytest.raises(ValueError, match="outside"),
         ):
             generate("test", output=evil)
@@ -113,8 +113,8 @@ class TestGenerateOutputValidation:
 
         # Act
         with (
-            patch("voicecli.config.load_defaults", return_value={}),
-            patch("voicecli.engine.get_engine", return_value=mock_engine),
+            patch("voicecli.core.config.load_defaults", return_value={}),
+            patch("voicecli.engines.engine.get_engine", return_value=mock_engine),
         ):
             result = generate(
                 "test", output=outside, allowed_base=UNRESTRICTED, engine="chatterbox"
@@ -138,8 +138,8 @@ class TestCloneOutputValidation:
 
         # Act + Assert
         with (
-            patch("voicecli.config.load_defaults", return_value={}),
-            patch("voicecli.engine.get_engine", side_effect=Exception("should not reach")),
+            patch("voicecli.core.config.load_defaults", return_value={}),
+            patch("voicecli.engines.engine.get_engine", side_effect=Exception("should not reach")),
             pytest.raises(ValueError, match="outside"),
         ):
             clone("test", ref=ref, output=evil)
@@ -156,8 +156,8 @@ class TestCloneOutputValidation:
 
         # Act
         with (
-            patch("voicecli.config.load_defaults", return_value={}),
-            patch("voicecli.engine.get_engine", return_value=mock_engine),
+            patch("voicecli.core.config.load_defaults", return_value={}),
+            patch("voicecli.engines.engine.get_engine", return_value=mock_engine),
         ):
             result = clone(
                 "test",
